feat: add marker to Map class
This commit adds a new method `addMarker` to the `Map` class. This method takes a `Mappable` object as a parameter and adds a marker to the Google Map at the specified location. The `Mappable` interface defines the structure of the object that can be passed to the `addMarker` method.

The `addMarker` method creates a new `google.maps.Marker` object and sets its position based on the `lat` and `lng` properties of the `Mappable` object. The marker is then added to the `googleMap` instance variable of the `Map` class.

This change allows markers to be added to the map for any object that implements the `Mappable` interface, such as the `User` and `Company` classes.