feat: add marker content to Mappable objects
This commit modifies the `Map` class and the `User` and `Company` classes to include a `markerContent` method. The `markerContent` method returns a string that represents the content to be displayed in the marker's info window.

By adding the `markerContent` method to the `Mappable` interface and implementing it in the `User` and `Company` classes, we can now customize the content of the info window for each marker on the map.

This change enhances the functionality of the map by providing more detailed information about the user or company associated with each marker.