feat: Implement Mappable interface in User and Company classes
This commit modifies the `User` and `Company` classes to implement the `Mappable` interface. By implementing this interface, the `User` and `Company` classes now have a `location` property with `lat` and `lng` coordinates.

This change is necessary to ensure that the `User` and `Company` objects can be passed as parameters to the `addMarker` method in the `Map` class, allowing markers to be added to the Google Map at the specified location.

This commit builds upon the previous commits that added marker functionality to the `Map` class and customized marker content for `Mappable` objects.